+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-05-02 17:10:20
-// Commit: feat(ui): implement responsive ui view component
+// Updated at: 2026-05-03 11:50:56
+// Commit: feat(ui): add bi-directional event bus handler for ui
 
 class uiComponentState {
-  final String commitRef = '793';
-  final String lastMessage = 'feat(ui): implement responsive ui view component';
-  final DateTime timestamp = DateTime.parse('2026-05-02T17:10:20+05:00');
+  final String commitRef = '810';
+  final String lastMessage = 'feat(ui): add bi-directional event bus handler for ui';
+  final DateTime timestamp = DateTime.parse('2026-05-03T11:50:56+05:00');
 }